Reorganized folder structure for features/base/fontIcons and used font icons instead of svg
