+++ features/base/fontIcons/components/_.native.js
@@ -1,0 +1,1 @@
+export * from './native';
