+++ features/base/fontIcons/components/index.js
@@ -1,0 +1,1 @@
+export * from './_';
