+++ features/base/fontIcons/components/native/index.js
@@ -1,0 +1,1 @@
+export * from './Icon';
