+++ features/base/fontIcons/components/web/index.js
@@ -1,0 +1,1 @@
+export * from './Icon';
