+++ features/base/fontIcons/components/web/styles/FontAwesome.css
@@ -6,16 +6,12 @@
  * -------------------------- */
 @font-face {
   font-family: 'FontAwesome';
-  /*
-   * FIXME features/base/styles shouldn't reference the files of
-   * features/base/fontIcons.
-   */
-  src: url('../../../../fontIcons/fonts/fontawesome-webfont.eot?v=4.6.3');
-  src: url('../../../../fontIcons/fonts/fontawesome-webfont.eot?#iefix&v=4.6.3') format('embedded-opentype'),
-       url('../../../../fontIcons/fonts/fontawesome-webfont.woff2?v=4.6.3') format('woff2'),
-       url('../../../../fontIcons/fonts/fontawesome-webfont.woff?v=4.6.3') format('woff'),
-       url('../../../../fontIcons/fonts/fontawesome-webfont.ttf?v=4.6.3') format('truetype'),
-       url('../../../../fontIcons/fonts/fontawesome-webfont.svg?v=4.6.3#fontawesomeregular') format('svg');
+  src: url('../../fonts/fontawesome-webfont.eot?v=4.6.3');
+  src: url('../../fonts/fontawesome-webfont.eot?#iefix&v=4.6.3') format('embedded-opentype'),
+       url('../../fonts/fontawesome-webfont.woff2?v=4.6.3') format('woff2'),
+       url('../../fonts/fontawesome-webfont.woff?v=4.6.3') format('woff'),
+       url('../../fonts/fontawesome-webfont.ttf?v=4.6.3') format('truetype'),
+       url('../../fonts/fontawesome-webfont.svg?v=4.6.3#fontawesomeregular') format('svg');
   font-weight: normal;
   font-style: normal;
 }
