+++ features/base/fontIcons/components/web/styles/index.js
@@ -1,0 +1,2 @@
+require('style!./FontAwesome.css');
+require('style!./jitsiFontIcons.css');
