+++ features/base/fontIcons/components/web/styles/jitsiFontIcons.css
@@ -1,0 +1,123 @@
+@font-face {
+    font-family: 'jitsi';
+    src:url('../../fonts/jitsi.eot?94d075');
+    src:url('../../fonts/jitsi.eot?#iefix94d075') format('embedded-opentype'),
+        url('../../fonts/jitsi.woff?94d075') format('woff'),
+        url('../../fonts/jitsi.ttf?94d075') format('truetype'),
+        url('../../fonts/jitsi.svg?94d075#jitsi') format('svg');
+    font-weight: normal;
+    font-style: normal;
+}
+
+[class^="icon-"], [class*=" icon-"] {
+    font-family: 'jitsi';
+    speak: none;
+    font-style: normal;
+    font-weight: normal;
+    font-variant: normal;
+    text-transform: none;
+    line-height: 0.75em;
+    font-size: 1.22em;
+
+    /* Better Font Rendering =========== */
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+}
+
+.icon-contactList:before {
+    content: "\e615";
+}
+.icon-avatar:before {
+    content: "\e616";
+}
+.icon-callRetro:before {
+    content: "\e611";
+}
+.icon-callModern:before {
+    content: "\e612";
+}
+.icon-recDisable:before {
+    content: "\e613";
+}
+.icon-recEnable:before {
+    content: "\e614";
+}
+.icon-authenticate:before {
+    content: "\e1ae";
+}
+.icon-kick1:before {
+    content: "\e60f";
+}
+.icon-kick:before {
+    content: "\e610";
+}
+.icon-share-desktop:before {
+	content: "\e602";
+}
+.icon-chat-simple:before {
+	content: "\e606";
+}
+.icon-full-screen:before {
+	content: "\e60d";
+}
+.icon-exit-full-screen:before {
+	content: "\e60e";
+}
+.icon-link:before {
+	content: "\e600";
+}
+.icon-chat:before {
+	content: "\e601";
+}
+.icon-presentation:before {
+	content: "\e603";
+}
+.icon-security:before {
+	content: "\e604";
+}
+.icon-share-doc:before {
+	content: "\e605";
+}
+.icon-telephone:before {
+    content: "\e611";
+}
+.icon-security-locked:before {
+	content: "\e607";
+}
+.icon-camera:before {
+	content: "\e608";
+}
+.icon-camera-disabled:before {
+	content: "\e609";
+}
+.icon-mic-disabled:before {
+	content: "\e60a";
+}
+.icon-microphone:before {
+	content: "\e60b";
+}
+
+.icon-hangup:before {
+    content: "\e617";
+}
+
+.icon-reload:before {
+    content: "\e618";
+}
+
+.icon-filmstrip:before {
+    content: "\e619";
+}
+
+.icon-connection:before {
+    line-height: normal;
+    content: "\e61a";
+}
+
+.icon-settings:before {
+    content: "\e61b";
+}
+
+.icon-dialPad:before {
+    content: "\e61c";
+}
