+++ features/base/fontIcons/constants.js
@@ -1,12 +1,12 @@
-import icoMoonConfig from './fonts/selection.json';
+import icoMoonConfig from './components/fonts/selection.json';
 
 /**
  * Lookup of Jitsi-specific Ico Moon icons.
  *
  * @type {Object}
  */
-export const JITSI_ICONS = icoMoonConfig.icons.reduce((memo, icon) => {
-    memo[icon.properties.name] = icon.icon.paths.join(' ');
+export const JITSI_ICONS = new Set(icoMoonConfig.icons.reduce((memo, icon) => {
+    memo.push(icon.properties.name);
 
     return memo;
-}, {});
+}, []));
