+++ features/base/fontIcons/index.js
@@ -1,1 +1,1 @@
-export * from './Icon';
+export * from './components';
