+++ features/base/styles/components/index.js
@@ -1,2 +1,1 @@
-export * from './_';
 export * from './styles';
