--- features/base/styles/components/web/index.js
@@ -1,1 +1,0 @@
-export * from './styles';
