--- features/base/styles/components/web/styles/index.js
@@ -1,1 +1,0 @@
-require('style!./FontAwesome.css');
